python/tracing/pipecat-openinference/requirements.txt
pipecat-ai[openai]
openinference-instrumentation-pipecat
openinference-instrumentation
respan-tracing
respan-instrumentation-openinference
python-dotenv
